+++ src/core/CHANGELOG.md
@@ -1,4 +1,7 @@
 # nstd_core
+## Latest
+- Added `nstd_core_str_from_cstring_with_null`.
+- The result of `nstd_core_str_from_cstring` no longer includes the null terminator.
 ## 0.3.2
 - `core.pointer` now uses `NSTDAny`.
 - Added `nstd_core_math_[degrees|radians]_f[32|64]`.